Modify code to use the for...loop
I wrote to a function for a single round then called the
function five times in another function. I called the new
function at the end of the code to have the game run five
rounds.

I modified it to use a for loop so as to add an ending
message after the game ends i.e after five rounds of play.