Update win message for each round
The win message used to be plain. You win, you lose, or
it's a tie. I want it to show the option that beat the
other so the win message will be in this format: You win!
A beats B.

I created two new variables, play and comp. I used the
slice method on the choice of each player to return what
they chose but with the first letter in uppercase. I then
concatenated the rest of the string length to the first
letter.

I also updated the conditions for the score count. I used
result === 'You win!' earlier. But the update I made to the
findWinner function renders the former === condition useless.

I used the .includes method to achieve what I wanted.