Updated code to count games won by each player
I created a new div to house the paragraph and span elements that will
display the scores after a game is won. The program adds a point to
the player's score if it find the word 'Congratulations' in the final
message.

In the same manner, it adds a point to the computer's score if it finds
the word 'Oops!' within the final message.